Added Supplementary Files
I decided to add the rest of the files that were in the old Gifty website. I will be editing them individually from here on out, with varying difficulty on each page. My plan is to continue to copy the Javascript to a separate file and adjust the algorithms accordingly. This way I can combine files at a later point if absolutely necessary.

I also added the moderation page, which was simply an afterthought with the old version. The idea was to be able to have a page where I could track all my users and their account's properties (especially since I was only planning on ever having a handful of users). If I ever plan on expanding my website to accept a larger amount of users, I would lkely adjust user accounts to be organized by family. Possibly more on that later.