Updated friendList Files
The biggest portion that was completed in this iteration was the friendList. Ensuring that everything was properly adapted to this page was a little rough, but I think it turned out well.

In addition, I also updated giftAddUpdate a little bit and gave both homeAlg and listsAlg a huge overhaul. I edited all of the critical pages in this update to use the same database calls. This increases simplicity, and hopefully less data usage later on as well.