Updated confirmation Files
Finished updating confirmation files. Whew! That took a while. I ended up using some ideas in this file to edit some other files. For example, I edited the databaseQuery to update the userDataArr every time the page is loaded. It's a little inefficient, but I'd rather act on data that I know is up to date all the time. In this particular case, the redundancy is okay.

The other updates here were minor updates to the html files, mostly just removing old comments that don't need to be there anymore.