Updated index Files
Whew! I've been working hard on updating all of the files as I scrutinize every detail to the best of my ability! For now, here's an update on indexAlg. One of the many updates (both for this file and the rest) was the removal of the reload timer, in exchange for a 15-minute logoff timer on the rest of the pages, that redirect the user here. Another update is more dynamic userData array storage. Any and all updates are stored as they happen, exactly where they need to be. To finish,  a minor addition, was the confirmation of the user's info in the "User [USER] Authenticated" notification.