Updated home Files
I updated the homeAlg file again with some updates to the loginTimer. I have confirmed that the loginTimer works as it is supposed to, so I added in some code that notifies the user that they have been inactive. It will notify them at 10 minutes and give them 5 minutes to respond before they are logged out.