Updated lists Files
Two birds with one stone! Rather, two files with one stone. In this update, I cleaned up (and added to) the listsAlg and friendListAlg files. For each of these, I have the loginTimer added in [almost] exactly as it is in the homeAlg.  Additionally, I added some code to help prevent unecessary database updates (gifts that are bought can't be marked as bought again, you can't un-buy gifts if you didn't buy it in the first place, etc) as well as some error checking with buyers that don't exist in the system anymore. Overall, this is turning out to be some very robust code, and I'm thoroughly enjoying what this project has become!