Updated All Alg Files/Added Supplemental ModeratorAlg Features
Whoops! I accidentally left out some code that handles the getCurrentUser() function on each page. If the JSON.parse() tries fetching a null pointer, it will fail and halt loading on the page. Normally it should just redirect to the login page, as the user is likely not logged in. Hopefully this will fix some issues, but I'll find out!

Additionally, I added supplemental features in the moderatorAlg files. This mostly just includes some more functionality for moderators, which allows moderators to give or take away permissions from other users.