Updated All Alg Files
Alright! This should complete the initial version of the private gift list! I want to work on fixing something extra with my friendList and privateFriendList algorithm files now...

Essentially, whenever an update comes through (for example, if a gift is bought), the modal should update, right? Wrong. I still haven't quite figured that out, so for the time being my fix will be to simply close the modal if that gift was updated. Of course, I wouldn't want to close the modal all the time, so that will be slightly trickier, but I'll find out!