Updated privateFriendList Files
Cleaned up the privateFriendList code a bit and finally implemented the currentModalOpen feature. I plan on adding this feature to the friendList code as well, but I will do that tomorrow.

Just as a side note, I will be taking a break from this project soon to work on another side project. I will likely take the break once friendList is updated.