Updated All Alg Files
Welp, in light of the notification update, I needed to shift around how some of the database operations as needed. Unfortunately I don't think all of these pages are as efficient as they used to be, with the tradeoff of having notifications for my handful of users on-demand.

Additionally, I finally added the err.toString() call on all of my getCurrentUser functions. That way if there does happen to be an error somewhere, I can figure out exactly why its erroring out.