Updated index Files
With this commit, I've completely overhauled the index page. Well, to be more accurate, I shuffled some stuff around. It now loads everything in a particular order AND you can now load in a config file of your choice! Just fill in the database info on /txt/config.txt and you'll be all set to go!

Another challenge that I hadn't thought about before was: "What if the user tries to sign up or login before the config is loaded?". Well, I fixed that problem, as the buttons now only work if the config and database have been properly initialized. Hopefully this should solve that problem!