Updated notifications Files
With this quick update, I have now added global notifications. Any users that are given a moderator role can now send a notification to all the users on the Gifty database. The most notable use would be to notify users of server downtime, but can be used for other important notifications as well.

Additionally, I did a little housekeeping and refactored a couple variable names, as well as added a second layer of moderatorAlg verification. This way, users that know about the moderator page can't access it unless they have the role.

Last but not least, I updated README to reflect current project status.