Updated All HTML Files *Breaking*
AS OF THIS POINT THE WEBSITE WILL NOT WORK PROPERLY DUE TO THE UPDATES IN THIS COMMIT. Once all JS files have been updated properly, this problem will be resolved. All HTML pages have been updated with new references to the lists. This allows for one unified reference among JS files.