Updated secretSantaAlg File
Updated secretSantaAlg JS file with overhauled Secret Santa name assignments. I've worked with the algorithm a bit more and I've been able to get upwards of a theoretical 75% success rate with worst-case considerations (Checking if the potential user match is the same as the current index user, checking friend list, checking family relationships, with only 4 users). I'm rather pleased with it thus far, but there's much more that I want to do! I want to shoot for 80 or 90% success, if possible.