Updated secretSantaAlg File
Updated secretSantaAlg JS file with a "shuffle mode" which forces certain button(s) to be disabled in the case that the function takes longer to process than expected. By doing so, the user is unable to kick off the process more than once.

Secret Santa name assignments have also been finer tuned with more detail in the console as to why a particular error may have occurred.