Updated homeAlg File
Updated homeAlg JS file with functionality for the new confirmModal, initialized gift tracking, new maintenance log ticket, removed key usage in gift add/update, and (finally) live gift deletion. For the latter, it would previously refresh the page to obtain the up-to-date data, and the processing has now been added for this (and will also be added elsewhere).